ISIS-2441: house keeping
